tests/test_authentication.py
"""Authentication tests for Huckleberry API."""

import asyncio
import time

import aiohttp
import pytest

from huckleberry_api import HuckleberryAPI


class TestAuthentication:
    """Test authentication functionality."""

    async def test_authenticate_success(self, api: HuckleberryAPI) -> None:
        """Test successful authentication."""
        assert api.id_token is not None
        assert api.refresh_token is not None
        assert api.user_uid is not None
        assert api.token_expires_at is not None

    async def test_authenticate_invalid_credentials(self, websession: aiohttp.ClientSession) -> None:
        """Test authentication with invalid credentials."""

        invalid_api = HuckleberryAPI(
            email="invalid@test.com", password="wrongpassword", timezone="UTC", websession=websession
        )
        with pytest.raises((RuntimeError, aiohttp.ClientResponseError, aiohttp.ClientError)):
            await invalid_api.authenticate()

    async def test_token_refresh(self, api: HuckleberryAPI) -> None:
        """Test token refresh functionality."""
        # Wait 1 second to ensure we get a new token (Firebase may return same token if too fresh)
        await asyncio.sleep(1)
        original_token = api.id_token
        await api.refresh_auth_token()
        assert api.id_token is not None
        assert api.id_token != original_token

    async def test_maintain_session(self, api: HuckleberryAPI) -> None:
        """Test maintain_session ensures token validity."""
        original_token = api.id_token
        original_expires = api.token_expires_at

        # Call maintain_session - shouldn't refresh if token is fresh
        await api.maintain_session()
        assert api.id_token == original_token
        assert api.token_expires_at == original_expires

        # Simulate expired token by setting expiry to past
        old_expiry = time.time() - 100
        api.token_expires_at = old_expiry

        # Now maintain_session should refresh
        await api.maintain_session()
        assert api.id_token is not None
        assert api.token_expires_at is not None

        # Verify token was actually refreshed by checking expiry was updated significantly
        # New expiry should be at least 3000 seconds in the future (Firebase tokens are ~1 hour)
        assert api.token_expires_at > time.time() + 3000, (
            f"Token expiry not properly refreshed: was {old_expiry}, now {api.token_expires_at}"
        )

        # Verify the refreshed token works by making a Firestore call
        children = await api.get_children()
        assert len(children) > 0


class TestChildrenRetrieval:
    """Test children data retrieval."""

    async def test_get_children(self, api: HuckleberryAPI) -> None:
        """Test retrieving children list."""
        children = await api.get_children()
        assert isinstance(children, list)
        assert len(children) > 0

        child_ids = [child.id_ for child in children if child.id_]
        assert len(child_ids) == len(set(child_ids))

        # Verify child data structure
        for child in children:
            assert child.id_
            assert child.childsName


class TestErrorHandling:
    """Test error handling."""

    async def test_operations_require_authentication(self, websession: aiohttp.ClientSession) -> None:
        """Test that operations fail without authentication."""
        import os

        email = os.getenv("HUCKLEBERRY_EMAIL", "test@example.com")
        password = os.getenv("HUCKLEBERRY_PASSWORD", "password")
        timezone = os.getenv("HUCKLEBERRY_TIMEZONE", "UTC")

        unauthenticated_api = HuckleberryAPI(email=email, password=password, timezone=timezone, websession=websession)

        # Note: API actually requires authentication but doesn't always raise
        # Firestore SDK may succeed with cached credentials from fixture
        # This test verifies the API can be instantiated without immediate auth
        assert unauthenticated_api.id_token is None

    async def test_invalid_child_uid(self, api: HuckleberryAPI) -> None:
        """Test operations with invalid child UID."""
        # Firebase Security Rules block writes to invalid child UIDs
        # This is expected - it will raise PermissionDenied
        from google.api_core.exceptions import PermissionDenied

        with pytest.raises(PermissionDenied):
            await api.start_sleep("invalid-uid-12345")
