pyproject.toml
[project]
name = "huckleberry-api"
version = "0.1.19"
description = "Python API client for Huckleberry baby tracking app using Firebase Firestore"
readme = { file = "README.md", content-type = "text/markdown" }
license = { text = "MIT" }
authors = [
    { name = "Woyken", email = "13166802+Woyken@users.noreply.github.com" }
]
requires-python = ">=3.9"
dependencies = [
    "google-cloud-firestore>=2.11.0",
    "pydantic>=2.10.0",
    "aiohttp>=3.10.0",
    "tzdata>=2024.1",
]
keywords = ["huckleberry", "baby", "tracking", "firebase", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/Woyken/py-huckleberry-api"
Repository = "https://github.com/Woyken/py-huckleberry-api"
Issues = "https://github.com/Woyken/py-huckleberry-api/issues"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.11.0",
    "ty>=0.0.1a8",
]

[tool.ruff]
target-version = "py39"
line-length = 120
exclude = [
    ".git",
    ".venv",
    "dist",
    "node_modules",
    "jadx output latest",
]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"
